+++ pages/index.tsx
@@ -1,5 +1,5 @@
 import Layout from '../components/layout/Layout'
-import { Courses } from '../components/Courses';
+import { Courses } from '../components/courses';
 import { NextPage } from 'next'
 
 const IndexPage: NextPage = () => {
